requirements.txt
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.11
coreapi==2.3.3
coreschema==0.0.4
crispy-bootstrap5==0.6
Django==4.0.1
django-crispy-forms==1.14.0
djangorestframework==3.13.1
drf-yasg==1.20.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
MarkupSafe==2.0.1
packaging==21.3
pkg_resources==0.0.0
psycopg2-binary==2.9.3
pyparsing==3.0.7
pytz==2021.3
requests==2.27.1
ruamel.yaml==0.17.20
ruamel.yaml.clib==0.2.6
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.8
